# mat64: optimize r/w of Dense and Vector