Fixes #346
```
name                    old time/op    new time/op    delta
MarshalDense10-4          2.66µs ± 1%    0.42µs ± 1%  -84.09%  (p=0.000 n=4+10)
MarshalDense100-4         22.3µs ± 1%     2.4µs ± 0%  -89.17%  (p=0.001 n=4+11)
MarshalDense1000-4         218µs ± 1%      23µs ± 1%  -89.54%  (p=0.001 n=4+11)
MarshalDense10000-4       2.18ms ± 1%    0.22ms ± 1%  -89.77%  (p=0.001 n=4+11)
UnmarshalDense10-4         918ns ± 0%     258ns ± 0%  -71.88%  (p=0.000 n=4+11)
UnmarshalDense100-4       3.83µs ± 0%    1.73µs ± 1%  -54.83%  (p=0.001 n=4+11)
UnmarshalDense1000-4      33.0µs ± 2%    16.0µs ± 1%  -51.33%  (p=0.001 n=4+11)
UnmarshalDense10000-4      309µs ± 0%     159µs ± 1%  -48.41%  (p=0.001 n=4+11)
MarshalVector10-4         2.48µs ± 1%    0.41µs ± 1%  -83.42%  (p=0.000 n=4+11)
MarshalVector100-4        21.9µs ± 1%     2.6µs ± 1%  -88.02%  (p=0.001 n=4+11)
MarshalVector1000-4        216µs ± 1%      24µs ± 1%  -88.66%  (p=0.001 n=4+11)
MarshalVector10000-4      2.15ms ± 1%    0.24ms ± 1%  -88.93%  (p=0.001 n=4+11)
UnmarshalVector10-4        811ns ± 0%     242ns ± 1%  -70.13%   (p=0.000 n=4+9)
UnmarshalVector100-4      3.75µs ± 0%    1.70µs ± 1%  -54.60%  (p=0.000 n=4+11)
UnmarshalVector1000-4     32.8µs ± 1%    16.0µs ± 1%  -51.17%  (p=0.001 n=4+11)
UnmarshalVector10000-4     309µs ± 1%     160µs ± 1%  -48.11%  (p=0.001 n=4+11)

name                    old alloc/op   new alloc/op   delta
MarshalDense10-4            400B ± 0%      208B ± 0%  -48.00%  (p=0.000 n=4+11)
MarshalDense100-4         2.64kB ± 0%    1.01kB ± 0%  -61.82%  (p=0.000 n=4+11)
MarshalDense1000-4        24.3kB ± 0%     8.3kB ± 0%  -65.88%  (p=0.000 n=4+11)
MarshalDense10000-4        242kB ± 0%      82kB ± 0%  -66.11%  (p=0.000 n=4+11)
UnmarshalDense10-4          272B ± 0%       80B ± 0%  -70.59%  (p=0.000 n=4+11)
UnmarshalDense100-4       1.90kB ± 0%    0.90kB ± 0%  -52.94%  (p=0.000 n=4+11)
UnmarshalDense1000-4      16.5kB ± 0%     8.2kB ± 0%  -50.34%  (p=0.000 n=4+11)
UnmarshalDense10000-4      164kB ± 0%      82kB ± 0%  -50.03%  (p=0.000 n=4+11)
MarshalVector10-4           384B ± 0%      208B ± 0%  -45.83%  (p=0.000 n=4+11)
MarshalVector100-4        2.62kB ± 0%    1.01kB ± 0%  -61.59%  (p=0.000 n=4+11)
MarshalVector1000-4       24.3kB ± 0%     8.3kB ± 0%  -65.86%  (p=0.000 n=4+11)
MarshalVector10000-4       242kB ± 0%      82kB ± 0%  -66.11%  (p=0.000 n=4+11)
UnmarshalVector10-4         256B ± 0%       80B ± 0%  -68.75%  (p=0.000 n=4+11)
UnmarshalVector100-4      1.89kB ± 0%    0.90kB ± 0%  -52.54%  (p=0.000 n=4+11)
UnmarshalVector1000-4     16.5kB ± 0%     8.2kB ± 0%  -50.29%  (p=0.000 n=4+11)
UnmarshalVector10000-4     164kB ± 0%      82kB ± 0%  -50.03%  (p=0.000 n=4+11)

name                    old allocs/op  new allocs/op  delta
MarshalDense10-4            26.0 ± 0%       2.0 ± 0%  -92.31%  (p=0.000 n=4+11)
MarshalDense100-4            206 ± 0%         2 ± 0%  -99.03%  (p=0.000 n=4+11)
MarshalDense1000-4         2.01k ± 0%     0.00k ± 0%  -99.90%  (p=0.000 n=4+11)
MarshalDense10000-4        20.0k ± 0%      0.0k ± 0%  -99.99%  (p=0.000 n=4+11)
UnmarshalDense10-4          8.00 ± 0%      1.00 ± 0%  -87.50%  (p=0.000 n=4+11)
UnmarshalDense100-4         8.00 ± 0%      1.00 ± 0%  -87.50%  (p=0.000 n=4+11)
UnmarshalDense1000-4        8.00 ± 0%      1.00 ± 0%  -87.50%  (p=0.000 n=4+11)
UnmarshalDense10000-4       8.00 ± 0%      1.00 ± 0%  -87.50%  (p=0.000 n=4+11)
MarshalVector10-4           24.0 ± 0%       2.0 ± 0%  -91.67%  (p=0.000 n=4+11)
MarshalVector100-4           204 ± 0%         2 ± 0%  -99.02%  (p=0.000 n=4+11)
MarshalVector1000-4        2.00k ± 0%     0.00k ± 0%  -99.90%  (p=0.000 n=4+11)
MarshalVector10000-4       20.0k ± 0%      0.0k ± 0%  -99.99%  (p=0.000 n=4+11)
UnmarshalVector10-4         6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=4+11)
UnmarshalVector100-4        6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=4+11)
UnmarshalVector1000-4       6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=4+11)
UnmarshalVector10000-4      6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=4+11)
```