@sbinet 2016-03-15
also added `io.EOF -> io.ErrUnexpectedEOF` for ncols,nrows.
no noticeable degradation wrt original improvement.